add theming platform utilities to the design package
