+++ packages/agentic-design/src/theming/index.ts
@@ -34,3 +34,4 @@ export { ThemeProvider } from './ThemeProvider';
 export type { ThemeProviderProps } from './ThemeProvider';
 export { ThemeReference } from './themeReference';
 export type { OnThemeChange, ThemeRecipe, ThemeTransform } from './themeReference';
+export { getCurrentAppearance, isHighContrast, setIsHighContrast } from './platformUtils';
