+++ packages/agentic-design/src/theming/platformUtils.ts
@@ -1,0 +1,1 @@
+export { getCurrentAppearance, isHighContrast, setIsHighContrast } from './platformUtils.defaults';
