+++ packages/agentic-design/src/theming/platformUtils.win32.ts
@@ -1,0 +1,7 @@
+import type { Theme } from './types/Theme.types';
+
+export { getCurrentAppearance, setIsHighContrast } from './platformUtils.defaults';
+
+export function isHighContrast(theme?: Theme): boolean {
+  return theme?.name === 'HighContrast';
+}
