+++ packages/theming/theming-utils/src/index.ts
@@ -1,4 +1,3 @@
-export { getCurrentAppearance } from './getCurrentAppearance';
-export { isHighContrast, setIsHighContrast } from './isHighContrast';
+export { getCurrentAppearance, isHighContrast, setIsHighContrast } from '@fluentui-react-native/design/theming';
 export { mapPipelineToTheme, mapFontPipelineToTheme } from './mapPipelineToTheme';
 export { mapPipelineToShadow } from './mapPipelineToShadow';
